+Gentoo support

+++ sysinfo.sh
@@ -72,6 +72,11 @@ if command_exists dnf ; then
     sudo dnf install -y gawk df nethogs sysstat lshw lsscsi hdparm dmidecode ; exit
 fi
 
+# For Gentoo
+if command_exists emerge ; then
+    sudo emerge --ask -y gawk df nethogs sysstat lshw lsscsi hdparm dmidecode ; exit
+fi
+
 # For RedHat / CentOS and derivatives
 if command_exists yum ; then
     sudo yum install -y gawk df nethogs sysstat lshw lsscsi hdparm dmidecode ; exit
